+++ app/src/components/ProfileDropDown/profileDropDown.scss
@@ -1,0 +1,8 @@
+
+.avatar {
+      background: url(../../../img/profile.svg);
+    height: 45px;
+    width: 45px;
+    background-size: 100%;
+    display: inline-block;
+}
